update OrderRepository to use Long ad ID

+++ order-service/src/main/java/org/tinkoff/orderservice/repository/OrderRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.tinkoff.orderservice.entity.Order;
 
-public interface OrderRepository extends JpaRepository<Order, Integer> , JpaSpecificationExecutor<Order> {
+public interface OrderRepository extends JpaRepository<Order, Long>, JpaSpecificationExecutor<Order> {
 }
